scrape/requirements.txt
 bs4
 psycopg2-binary
 pandas
 requests
